refactor(lib::playlist): asx tests to asx module
